autoNews/requirements.txt
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0
APScheduler==3.10.4
PyYAML==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.2.0
sumy==0.11.0
tqdm==4.66.1
click==8.1.7
python-dateutil==2.8.2
pytest==7.4.4

# TTS (Text-to-Speech)
edge-tts==7.2.7

# AI Integration
openai>=2.0.0
